Add Entity Framework packages
I decided to add the NuGet-packages related to EF Core in the Domain-project. This is because I usually work with a structure more similar to "clean architecture", but this is beyond the scope of this workshop. In short, Entity Framework is a dependency I would like to hide in an Infrastructure-layer.

The API-project requires the design-package to create migrations, but we will fix that with a DesignTimeFactory in the next commit.